Publisher#flatMapSingle make subscription non-volatile (#1025)
Motivation:
Publisher#flatMapSingle has the Subscription from onSubscribe saved to a
volatile variable. This was due to the Subscription being in the context
of the mapped Single's Subscriber termination callbacks and ambiguity in
the Reactive Streams specification about visibility related to
Publisher#subscribe(..) and Subscriber state. However recent
[discussions](https://github.com/reactive-streams/reactive-streams-jvm/issues/486)
have provided more insight that the Publisher must provide visibility,
and therefore the volatile state is not necessary.

Modifications:
- Publisher#flatMapSingle subscription member variable can be
non-volatile

Result:
Less volatile state and more clear expecations.