`SingleConcatWithPublisher` and invalid request-n
__Motivation__

If invalid calls to request-n are made for `Single.concat(Publisher)` before the `Single` result is propagated, we first emit the result and then propagate the invalid demand.
For cases, when the concatenated `Publisher` completes inline with `subscribe()` (eg: `Publisher.empty()`), this means we will never deliver an error for the invalid demand.

__Modification__

Deliver error when invalid demand is made in the above case instead of `Single`'s result and do not subscribe to the concatenated `Publisher`

__Result__

Invalid demand always terminates the `Subscriber` with an error.