Disable caching of CNAMEs to workaround resolving issues
Motivation:

In netty there is a seperate CNAME cache that is used to cache CNAMEs that are resolved during resolving an A / AAAA record. While this sounded like a great idea when I first implemented this in netty it turned out this is problematic and may produce errors while resolving hostnames later on.
The problem with the CNAME cache is that we may end up with entries in the CNAME cache wile the original cached resolution was already removed from the cache (due TTL). In this case we may find an entry in teh CNAME cache that we will then try to resolve via the specified nameserver. This resolution may fail as the nameserver may not return an entry for the previous cached CNAME. This usually happens when we receive the full "resolution chain" during query for an A / AAAA record.

Modifications:

Introduce NoopDnsCnameCache and explicit set it when building the used DnsNameResolver. This is a workaround as netty itself will do the same in the next release and deprecate DnsCnameCache

Result:

No more resolution failures caused by cached CNAME entries