Replace MaxSizeBasedWriteDemandEstimator with EWMAWriteDemandEstimator
Motivation:
MaxSizeBasedWriteDemandEstimator allocates an array of 32 elements by
default to keep historical size information. This can introduce a lot of
allocation and has been shown in benchmarks to consist of >3% of
allocations inside a TLAB and >%90 of allocations outside of a TLAB. The
estimation fidelity can be reduced with an Exponential Weighted Moving
Average given the loose estimation requirements which can also reduce
the allocation pressure.

Modification:
- Remove MaxSizeBasedWriteDemandEstimator
- Add EWMAWriteDemandEstimator which uses Exponential Weighted Moving
Average to estimate requestN

Result:
Less memory pressure introduces while estimating requestN for writing
data.