Do not shutdown output channel until client receives the full response (#1090)
Motivation:

When client observes "Connection: close" header it shutdowns the output
channel as soon as request is written. Some servers interpret FIN from
the client as an indicator that it lost interest in their data and
therefore, server just closes the second half of the connection asap.
As the result, connection may be closed before client receives the full
response from the server.

Modifications:

- Reproduce this scenario using a simple proxy tunnel that is not aware
of HTTP protocol semantics;
- Defer connection closure on the client-side until it completes the
full request-response iteration;
- Fail all subsequent or pipelined requests on the connection that moves
to the "closing" state;
- Adjust `RequestResponseCloseHandlerTest` to not expecting outbound
half-closure on the client side;
- Add more tests to verify that client handles "Connection: close" header
correctly;

Result:

Response is not aborted when client observes "Connection: close" header.