Configure `TransportObserver` via `ConnectionFactory` on the client-side (#1114)
Motivation:

Having `TransportObserver` on `ConnectionFactory#newConnection` method gives
users more flexibility when they need to correlate prior events and a state
with observability events of new connections. It also allows them to choose
for which `ResolvedAddress`es they need an observer.

Modification:

- Add `TransportObserver` as a second argument for
`ConnectionFactory#newConnection`;
- Remove `#transportObserver(...)` configuration method from all client builders;
- Introduce `TransportObserverConnectionFactoryFilter` to simplify `TransportObserver`
configuration for the client;
- Update code to use new `ConnectionFactory` API;

Results:

More flexible approach for users to configure `TransportObserver` on the
client-side.