Introduce `CatchAllTransportObserver` that catches and logs unexpected exceptions
Motivation:

We have to make sure that all invocations of observer methods are safe and do not
impact normal ST operations. Currently, we do that via `safeReport` utilities that
are used everywhere. It's easy to forget to make an invocation safe in all places.

Modifications:

- Add `CatchAllTransportObserver` implementation that makes sure all unexpected
exceptions are logged;
- Wrap the passed `TransportObserver` with `CatchAllTransportObserver` instead of
using `safeReport` utilities for all invocations of observer API;

Result:

Single places that makes sure all calls to `TransportObserver` are safe and do not
throw an exception.