Introduce `CatchAllTransportObserver` that catches and logs unexpected exceptions (#1118)
Motivation:

We have to make sure that all invocations of observer methods are safe and do not
impact normal ST operations. Currently, we do that via `safeReport` utilities that
are used everywhere. It's easy to forget to make an invocation safe in all places.

Modifications:

- Add `CatchAllTransportObserver` implementation that makes sure all unexpected
exceptions are logged;
- Add public `TransportObservers` factory that can convert any observer into its safe
version;
- Wrap the passed by users `TransportObserver` with
`TransportObservers#asSafeObserver` instead of using `safeReport` utilities for all
invocations of observer API;

Result:

A single place that makes sure all calls to `TransportObserver` are safe and do not
throw an exception.