Introduce HTTP content encoding H1 & H2
Motivation:

Our HTTP clients & servers don't support any form of compression. There are many use cases that involve large payloads, that compression could make a difference on the HTTP timings.

Modifications:

H1 & H2 protocols were altered to support a list of supported-encodings. We provide the default ones out of the box as part of the helper ContentEncodings.all() which offers, GZIP & DEFLATE.
To keep the behavior consistent between the two protocols, we chose to communicate the compression technique to the relevant endpoints through the 'Content-Encoding & Accept-Content-Encoding' headers, rather than on the 'Transfer-Encoding & TE' headers, which are not supported by H2.

Result:

Once the client & server are configured to support different encodings through the protocol configuration, a client can encode a request by passing the encoding request.encoding(...) whereas the server will decide on the encoding based on the client advertised headers. To disable the encoding for a particular request, or enforce a specific encoding, one can call response.encoding(...) to do so. That, bypasses the auto-selection, and also ignores compatibility checks with the client.

Encoding & Decoding is done in a streaming fashion.