Avoid `Connection reset by peer` error when server closes the connection
Motivation:

When users close the server gracefully or when the server adds a
`Connection: close` header to the response, the `Channel` will be closed as
soon as server read the request and wrote the response. Because a client is
not aware that the server intends to close the connection, it may send a
following request on the same connection before it reads the response.
In this case, TCP stack on the server-side will respond with RST frame
(because the `Channel` is already closed) that may erase data on the
connection that were delivered to the client but not acknowledged.
See https://tools.ietf.org/html/rfc7230#section-6.6 for more information.

Modifications:

- When the request is read swap the `HttpRequestDecoder` with a handler
that discards all new incoming requests;
- When response is written half-close the output side of the connection;
- When the FIN is received from the client, close the `Channel`;
- For SSL connections send `close_notify` before `shutdownOutput()`;
- Save the original `CloseEvent` and use it later to produce more accurate
logs for connection closure;
- Update `RequestResponseCloseHandlerTest` to account for a new state
machine;
- Enhance `GracefulConnectionClosureHandlingTest` to test the same
scenarios when graceful closure is initiated on the server-side;
- Add `ServerGracefulConnectionClosureHandlingTest` that reproduces
`Connection reset by peer` issue described in `Motivation` section;
- Minor improvements for `ConnectionCloseHeaderHandlingTest` to align it
with other tests;

Result:

Server does to cause `Connection reset by peer` exception when it closes
gracefully or adds `Connection: close` header.