Support batched service discoverer updates (#1146)
__Motivation__

`SingleAddressHttpClientBuilder` is used as a template for partitioned and multi-address builders for which SD event types is different. Although, users are not expected to use a ServiceDiscoverer with different event type, we accommodate this usage in our public APIs. Accomodating this usage further trickles down to builder implementation and the external facing `ServiceDiscoveryRetryStrategy`. It also complicates generic handling whenever we change `ServiceDiscoverer` APIs.
Further the partitioned builder uncheck casts the service discovery stream while using which hides type issues.

__Modification__

 - Remove the `? extends` in the `ServiceDiscoverer` event type while accepting a `ServiceDiscoverer` in the builders. When a different event type is required (partitioned builder), store the `ServiceDiscoverer` locally.
 - Simplify `ServiceDiscoveryRetryStrategy` generics.

 __Result__

 Less generics clutter in public APIs.