Initial support for gRPC message-encoding (#1151)
__Motivation:__

Add support for gRPC message-encoding according to https://github.com/grpc/grpc/blob/master/doc/PROTOCOL-HTTP2.md

__Modifications:__

Both Server & Client have changes that allow end-users to provide supported-encodings for each side. The granularity of the settings are, on the server side per service and on the client side per client.
The compression of each individual request is decided for each call separately with the use of the metadata, and for each response is decided automatically based on the server & client configurations.
Further improvements on this area, can allow in the future to control the supported-encodings from the client side per call, thus finer control over what encoding the response will use. This is not implemented in this patch.

__Result:__

gRPC services & clients can now use default or custom message encodings.