Single#toCompletionStage() avoid cancel if normal termination
Motivation:
Single#toCompletionStage() will cancel unconditionally if the Single
terminates normally or if the CompletableFuture is manually completed.
Cancellation is only necessary in the latter case and can create
additional work and complicates debugging in the former.

Modifications:
- Make a best effort to avoid cancellation if the Single terminates via
onSuccess or onError.
- These changes are also inherited by other sources (Publisher,
Completable) which leverage Single's conversion to CompletionStage.

Result:
Single, Completable, Publisher conversion to CompletionStage (and
CompletableFuture) no longer cancel when terminated normally.