Publisher#flatMapMerge queue not drained bug
Motivation:
Publisher#flatMapMerge may not drain the queue if an item is enqueued
and the item doesn't require demand. This may happen for example if the
mapped subscriber exhausts its demand, and had signals previously
queued. This may lead to dead lock and lack of progress.

Modifications:
- Modify tryEmitItem to only lock for item delivery. For other cases
that go through the queue use enqueueAndDrain which will enqueue,
acquire the drain lock, and drain the queue.

Results:
No more deadlock and smaller lock critical section in tryEmitItem for
Publisher#flatMapMerge.