Do not complete server write if there are still pending requests (#1155)
Motivation:

`RequestResponseCloseHandler.protocolPayloadEndOutbound` callback triggers
`ProtocolPayloadEndEvent` when server is in closing state without accounting
for pending requests. As the result, server will not send a response for the
second pipelined request, will not transition to the idle state, and will never
complete close the connection.

Modifications:

- Account for `pending` value before emitting `ProtocolPayloadEndEvent`;
- Renamve `ProtocolPayloadEndEvent` -> `OutboundDataEndEvent`;
- Add a test to verify server does not trigger `OutboundDataEndEvent`
while requests are pending;
- Add more tests to verify that `PROTOCOL_CLOSING_INBOUND`,
`PROTOCOL_CLOSING_OUTBOUND`, and `USER_CLOSING` events are
correctly handled for pipelined server connection;

Result:

Server responds to pending requests and closes the connection if it's already
in closing state while 2+ pipelined requests are in process.