Http URI related fixes and enhancements
Motivation:
Jersey's ContainerRequest requires that an API be encoded properly
before processing. The user agent is expected to do the encoding [1] but
not all user agents do, so we attempt to encode on the server for.
convenience. However the method we used to encode the URI is an internal
method to Jersey and doesn't cover the full scope of encoding defined in
rfc3986 [1]. The JDK's URI class also targets an obsolete rfc2396 [2]
and the JDK's URLEncoder class provides
application/x-www-form-urlencoded which is also not what we need.

Modifications:
- Modify HttpUri to validate RFC 3986 semantics, extract out the host
extraction from headers code outside the scope of this class. This
simplifies the logic within the Uri implementation and makes it easier
to validate against the RFC.
- Introduce an implementation of Uri that targets authority-form URIs
targeted specifically for CONNECT requests [3].
- Introduce 3 new filters. 1 to encode the request target on the client.
1 to decode the request target on the server. 1 to encode the request
target on the server (which can be used to replicate existing behavior
where Jersey is feed an encoded URI).
- Update DefaultHttpRequestMetaData so that effective host/port doesn't
become stale if the port value in the host header changes.

Result:
Uri handling follows the RFC [1] and encoding/decoding can be optionally
applied via a filter if desired.

[1] https://tools.ietf.org/html/rfc3986
[2] https://tools.ietf.org/html/rfc2396
[3] https://tools.ietf.org/html/rfc7230#section-5.3.3