Option to configure capacity of the datagram packet buffer for DNS SD (#1193)
Motivation:

UDP packet may be truncated if it exceeds the maximum payload size of 4096
bytes. In this case, DNS ServiceDiscoverer will retry with TCP. However,
users should be able to configure higher capacity of the datagram packet
buffer.

Modifications:

- Add `DefaultDnsServiceDiscovererBuilder#maxUdpPayloadSize` method;
- Fix `IllegalArgumentException` message for `minTTL` option;

Result:

Users can control the capacity of the datagram packet buffer.