Publisher#flatMapMerge recover from mapped Publisher exception (#1218)
Motivation:
Publisher#flatMapMerge does not attempt to recover from throwing in
onNext from a mapped publisher. In this case we should attempt to
replenish demand incase there is a recoverWith or similar operator
applied to the mapped publisher, or else we may deadlock.

Modifications:
- FlatMapSubscriber should try to replenish demand even if onNext throws
an exception

Result:
Publisher#flatMapMerge is compatible with recoverWith style operators on
the mapped publishers.