Improve exception messages from `HttpObjectDecoder`
Motivation:

`HttpObjectDecoder` and its subclasses may throw an exception if the received
HTTP message violates the RFC and can not be parsed. However, exception messages
are short and hard to understand, without any pointers to what part of the
message caused an exception. Those cases are hard to debug and usually require
knowledge of RFC.

Modifications:

- Introduce `IllegalCharacterException` type that shows an illegal character as
a `char` and in hex format, as well as an expected values;
- `try-catch` an `IllegalCharacterException` and wrap it with `DecoderException`
that clarifies where the illegal character was found;
- Adjust tests for new exception messages;

Result:

More understandable exception messages from `HttpObjectDecoder`.