DNS Client SRV robustness and A* record Publisher support (#1223)
Motivation:
SRV resolution involves a two phase lookup. First a SRV query, which
returns a set of name/port pairs, and then for each pair a A* record
resolution. If an A* record resolution fails it currently fails the
outer SRV Publisher. We also do a one-time-lookup at the A* record level
so we can't detect multiple A* records nor can we detect any changes at
the A* record level. When a failure at the SRV level occurs we don't
cleanup the A* publisher query processing, which may accumulate queued
items over subsequent retry attempts.

Modifications:
- DefaultDnsClient SRV resolution should use flatMapMerge to process
multiple A* resolution streams in parallel, track changes, and deal with
dynamic failures.
- DefaultDnsClient SRV resolution should not let A* resolution failures
propagate to the SRV level, and instead should be retried as long as
there is a corresponding SRV entry.
- DefaultDnsClient SRV resolution should force innactive events upon
failure so that A* record publishers cleanup associated queues/timers.
- DefaultDnsClient SRV resolution should support a mode that allows
deduplication of <host, ip> pairs returned by different A* record
publishers.
- DefaultDnsClient SRV resolution should disable Netty's CNAME cache
because it may result in incorrect failures if CNAME TTLs are longer
than the SRV entries TTLs they map to.
- DefaultDnsClientTest and TestRecordStore have been reimplemented to
allow for direct control in the unit test, rather than implicit
removal/addition in the background.

Result:
More robust DNS SRV record resolution that can also continiously follow
each corresponding A* record stream.