Github Actions, use pull_request and workflow_run instead of pull_request_target
Motivation:
pull_request_target will run against the commit where a PR is branched
off of. This may result in errors not being detected until after the
change is merged (e.g. checkstyle, etc.).

Modifications:
- Instead use the pull_request event to build pull requests, and publish
the artifacts which need post analysis (test results, checkstyle, etc.).
A workflow_run job can run with the permissions of the original repo and
post annotations to PRs indicating any issues with the build. This
separation is done for security reasons to prevent forks from getting
access to credentials/secrets.

Result:
Pull requests build/analysis is done on the code in the pull request,
instead of from the target branch.