DynamicCompositeCancellable race condition fix
Motivation:
DynamicCompositeCancellable are used in operators that keep a collection
of Cancellable objects (e.g. flatMap, merge, etc.). The implementations
are intended to be thread safe between adding Cancellables and
cancelling Cancellables that have been previously add (and not removed).
The DynamicCompositeCancellable implementations rely upon
reading volatile state after an element is added to a concurrent
collection to cancel the new element if there is a race. However the add
thread is not gaurenteed to observe a write from the cancel thread, and
if this is the case an element maybe added and never cancelled.

Modifications:
- The add thread should attempt to read/modify state in coordination
with the cancel thread to gaurentee that element will be cancelled.

Result:
More correct concurrency code in DynamicCompositeCancellable.