Handle `close_notify` TLS alert (#1285)
Motivation:

TLS protocol defined `close_notify` alert message [1], which notifies that the
sender will not send any more messages on this connection. `SslHandler`
generates `SslCloseCompletionEvent` in this case, then transport emits
`ChannelInputShutdownReadComplete` event when the FIN is received. There is a
race between these 2 user events and sending a new request on the client-side.
We MUST close down the connection immediately, discarding any pending writes,
and preventing new requests. We should also mark that the connection is
closing asap to prevent LB from selecting this connection.

1. https://tools.ietf.org/html/rfc5246#section-7.2.1

Modifications:

- Handle `SslCloseCompletionEvent` in `DefaultNettyConnection`;
- Add tests to verify client and server behavior when `SslCloseCompletionEvent`
is received;

Result:

The connection transitions to the "closing" state as soon as the
`SslCloseCompletionEvent` is observed, preventing new writes on this connection
and failing the first write, making failures retryable.