`DefaultSingleAddressHttpClient` may add incorrect `host` header by default
Motivation:

Existing approach for adding `HostHeaderHttpRequesterFilter` in
`DefaultSingleAddressHttpClientBuilder` keeps reference to the
`hostToCharSequenceFunction` at the time of class initialization. However,
`hostToCharSequenceFunction` can be changed later (in a proxy ctor or
`unresolvedAddressToHost` builder method). In these cases,
`hostHeaderFilterFactoryFunction` keeps the reference to the old
`hostToCharSequenceFunction`.

Modifications:

- Use a `boolean` to know when `HostHeaderHttpRequesterFilter` should be
applied;
- When `HostHeaderHttpRequesterFilter` is added, use the latest reference to
the `hostToCharSequenceFunction`;

Result:

`HostHeaderHttpRequesterFilter` uses the latest version of
`hostToCharSequenceFunction` to determine the fallback value for the `host`
header.