HTTP Trailers Propagation Fix
Motivation:
The HTTP protocol represents data as a single flat stream of data of the
following format {MetaData+Headers, Data*, Trailers*}. Trailers may or
may not be present depending when reading from the transport, and users
don't always add trailers. Because of this we attempt to avoid including
trailers in the stream unless we read them from the transport or the
user explicitly does a transformation to add trailers. This helps reduce
the cost/complexity of intermediate transformations which may need to
expose the Buffers and still preserve the trailers. However this
conditional logic doesn't reliably work on the client side because the
control flow merges the read and write stream and we only concatenate
the trailers after the response stream completes successfully, which
maybe too late if the write stream never completes (or only completes in
error when the transport is closed). The end result is trailers may be
dropped even when they are explicitly added (e.g. in gRPC).

Modifications:
- Rework StreamingHttpPayloadHolder to use a merge operation so the
  trailers are processed inline with the flat stream, and don't depend
  upon the stream completion.
- StreamingHttpPayloadHolder to avoid storing state outside the
  Publisher transformations, and instead state should be coupled with
  each subscribe operation to be more ReactiveStreams compliant and
  supporting use cases like re-subscribe.
- Deprecate "payloadAndTrailers()" and "raw" transformation methods. The
  raw payload type was intended to be used for FileRegion but this is
  work is currently unplanned and memory mapped files as Buffers is the
  current preferred alternative (which also works with TLS is enabled).
  These methods will be removed in a future commit and this will also
  simplify the internal implementation of StreamingHttpPayloadHolder and
  other areas that deal with the underlying payload.

Result:
Trailers are preserved on the client and server side regardless of the
payload stream completion status.