Make `ServiceTalkWireLogger` consistent with `LoggingHandler` (#1309)
Motivation:

`ServiceTalkWireLogger` is a rework if netty's `LoggingHandler` which allows
us to control logging of user data. However, the output format was a bit
different from netty, making it harder to compare netty and ST logs. Also,
it has a few minor issues with the output.

Modifications:

- Align output format with `LoggingHandler`;
- Make internal methods `static` when possible;
- Do not print extra new line when the buffer is empty;
- Log how `Channel`'s writability changed;
- Do not log `localAddress` when it's `null`;

Result:

Output of `ServiceTalkWireLogger` and `LoggingHandler` is aligned.