Some filters won't be executed on retry/repeat (#1302)
Motivation:

`HostHeaderHttpRequesterFilter`, `AbsoluteAddressHttpRequesterFilter`,
and `AbstractStreamingHttpConnection` modify the request outside of the
asynchronous execution chain. In case of repeat or retry, if the
`request`'s state was reset or a new `request` was created, we loose
the state applied by those classes.

Modifications:

- Use `Single.defer` operator in all these cases to execute the logic
for each retry/repeat operation;
- Updates tests when necessary;

Result:

Filters logic always applied on each retry/repeat.