Make `AsyncContext` visible from `TransportObserver#onNewConnection()` (#1303)
Motivation:

Users who want to correlate `TransportObserver#onNewConnection()` events with
the request object may use `AsyncContext` to propagate correlation key.
However, `AsyncContext` is not visible from `TransportObserver#onNewConnection()`
because it runs from IO-thread. It's also important to invoke
`TransportObserver#onNewConnection()` callback earlier on the client-side.

Modifications:

- Trigger `TransportObserver#onNewConnection()` as soon as someone subscribes to
`TcpConnector#connect` single on the client-side to get access to the caller's
thread `AsyncContext`;
- Trigger `TransportObserver#onNewConnection()` at the beginning of `initChannel`
on the server-side for consistency with the client-side;
- Update internal classes and tests to use new API;
- Add tests to verify `AsyncContext` is visible for expected callbacks;

Result:

`AsyncContext` is visible from `TransportObserver#onNewConnection()` on the
client-side.