Don't close entire HTTP/2 connection when request is cancelled (#1307)
Motivation:

`LoadBalancedStreamingHttpClient` pessimistically assumes the connection
that serves a request should be closed on request cancellation. However,
in case of HTTP/2 it closes the parent TCP connection instead of a stream
(because stream is not accessible from the connection API). It also fails
all other in-flight requests on the same connection.

Modifications:

- In case of HTTP/2 and above mark request as finished instead of closing
the connection when cancellation happens;
- Test that request cancellation doesn't close parent TCP connection for
HTTP/2 protocol;
- Change `HttpConnectionContext#protocol()` to return `HttpProtocolVersion`;
- Change `GrpcConnectionContext.GrpcProtocol` to return `HttpProtocolVersion`;

Result:

HTTP/2 request cancellation doesn't close the TCP connection.