Publisher#flatMapMerge* preserve error
Motivation:
The Publisher#flatMapMerge* operators call cancel() on the upstream
subscription before calling onError(..) on the downstream subscriber in
the event of errors. This may result in cleanup without knowledge of the
root cause and even result in incorrect status to be propagated (e.g.
error is recovered and translated into a response).

Modifications:
- Publisher#flatMapMerge* implementations attempt to propagate
  onError(..) before cancel upstream subscription
- Remove the atomic terminalNotification from Publisher#flatMapMerge*
  implementaitons as it is no longer necessary becuase we keep a fine
  grained count of how many mapped sources are outstanding which lets us
  know exactly when we have terminate.
- Remove CompositeException which adds additional layer of queue and
  transfer to suppressed exception list in THrowable. This also
  throws an exception type which must be unwrapped in order to
  understand the root cause. Change default number of suppressed
  exceptions to 0 by default to avoid potential memory issues if static
  exceptions are thrown.

Result:
More informative errors when Publisher#flatMapMerge*(..) is used.